move next up list data loading
