+++ src/utils/fetcher.ts
@@ -1,0 +1,2 @@
+const fetcher = (url: RequestInfo) => fetch(url).then((r) => r.json())
+export default fetcher
